+++ packages/core/components/text-editor/text-editor-class.js
@@ -7,7 +7,7 @@ const textEditorButtonsMap = {
   // f7-icon, material-icon, command
   bold: ['bold', 'format_bold', 'bold'],
   italic: ['italic', 'format_italic', 'italic'],
-  underline: ['underline', 'format_underline', 'underline'],
+  underline: ['underline', 'format_underlined', 'underline'],
   strikeThrough: ['strikethrough', 'strikethrough_s', 'strikeThrough'],
   orderedList: ['list_number', 'format_list_numbered', 'insertOrderedList'],
   unorderedList: ['list_bullet', 'format_list_bulleted', 'insertUnorderedList'],