+++ packages/react/components/input.d.ts
@@ -7,6 +7,7 @@ declare namespace F7Input {
     name? : string
     value? : string | number | Array<any> | Date | Object
     defaultValue? : string | number | Array<any>
+    inputmode? : string
     placeholder? : string
     id? : string | number
     className? : string