+++ packages/react/components/list-item.d.ts
@@ -18,6 +18,8 @@ declare namespace F7ListItem {
     tooltipTrigger? : string
     link? : boolean | string
     target? : string
+    tabLink? : boolean | string
+    tabLinkActive? : boolean
     after? : string | number
     badge? : string | number
     badgeColor? : string