+++ packages/svelte/components/list-input.svelte
@@ -25,6 +25,7 @@
   export let type = 'text';
   export let name = undefined;
   export let value = undefined;
+  export let inputmode = undefined;
   export let readonly = undefined;
   export let required = undefined;
   export let disabled = undefined;
@@ -429,6 +430,7 @@
                 placeholder={placeholder}
                 id={inputId}
                 size={size}
+                inputmode={inputmode}
                 accept={accept}
                 autocomplete={autocomplete}
                 autocorrect={autocorrect}
@@ -475,6 +477,7 @@
                 style={inputStyle}
                 name={name}
                 type={inputType}
+                inputmode={inputmode}
                 placeholder={placeholder}
                 id={inputId}
                 size={size}
@@ -510,7 +513,7 @@
             {/if}
           {/if}
           <slot name="input" />
-          {#if hasErrorMessage && errorMessageForce} && (
+          {#if hasErrorMessage && errorMessageForce}
             <div class="item-input-error-message">
               {Utils.text(errorMessage)}
               <slot name="error-message"/>
@@ -610,6 +613,7 @@
               placeholder={placeholder}
               id={inputId}
               size={size}
+              inputmode={inputmode}
               accept={accept}
               autocomplete={autocomplete}
               autocorrect={autocorrect}
@@ -656,6 +660,7 @@
               style={inputStyle}
               name={name}
               type={inputType}
+              inputmode={inputmode}
               placeholder={placeholder}
               id={inputId}
               size={size}
@@ -691,7 +696,7 @@
           {/if}
         {/if}
         <slot name="input" />
-        {#if hasErrorMessage && errorMessageForce} && (
+        {#if hasErrorMessage && errorMessageForce}
           <div class="item-input-error-message">
             {Utils.text(errorMessage)}
             <slot name="error-message"/>