+++ packages/vue/components/list-item.js
@@ -19,6 +19,8 @@ export default {
     tooltipTrigger: String,
     link: [Boolean, String],
     target: String,
+    tabLink: [Boolean, String],
+    tabLinkActive: Boolean,
     after: [String, Number],
     badge: [String, Number],
     badgeColor: String,
@@ -89,6 +91,8 @@ export default {
       header,
       footer,
       link,
+      tabLink,
+      tabLinkActive,
       href,
       target,
       after,
@@ -161,11 +165,14 @@ export default {
       if (link || href || accordionItem || smartSelect) {
         const linkAttrs = Object.assign({
           href: link === true ? '' : link || href,
-          target
+          target,
+          'data-tab': Utils.isStringProp(tabLink) && tabLink || undefined
         }, Mixins.linkRouterAttrs(props), {}, Mixins.linkActionsAttrs(props));
         const linkClasses = Utils.classNames({
           'item-link': true,
-          'smart-select': smartSelect
+          'smart-select': smartSelect,
+          'tab-link': tabLink || tabLink === '',
+          'tab-link-active': tabLinkActive
         }, Mixins.linkRouterClasses(props), Mixins.linkActionsClasses(props));
         linkEl = _h('a', __vueComponentTransformJSXProps(Object.assign({
           ref: 'linkEl',