+++ packages/vue/components/page.js
@@ -231,7 +231,7 @@ export default {
   },
 
   created() {
-    Utils.bindMethods(this, ['onPtrPullStart', 'onPtrPullMove', 'onPtrPullEnd', 'onPtrRefresh', 'onPtrDone', 'onInfinite', 'onPageMounted', 'onPageInit', 'onPageReinit', 'onPageBeforeIn', 'onPageBeforeOut', 'onPageAfterOut', 'onPageAfterIn', 'onPageBeforeRemove', 'onPageBeforeUnmount', 'onPageStack', 'onPageUnstack', 'onPagePosition', 'onPageRole', 'onPageMasterStack', 'onPageMasterUnstack', 'onPageNavbarLargeCollapsed', 'onPageNavbarLargeExpanded', 'onCardOpened', 'onCardClose']);
+    Utils.bindMethods(this, ['onPtrPullStart', 'onPtrPullMove', 'onPtrPullEnd', 'onPtrRefresh', 'onPtrDone', 'onInfinite', 'onPageMounted', 'onPageInit', 'onPageReinit', 'onPageBeforeIn', 'onPageBeforeOut', 'onPageAfterOut', 'onPageAfterIn', 'onPageBeforeRemove', 'onPageBeforeUnmount', 'onPageStack', 'onPageUnstack', 'onPagePosition', 'onPageRole', 'onPageMasterStack', 'onPageMasterUnstack', 'onPageNavbarLargeCollapsed', 'onPageNavbarLargeExpanded', 'onCardOpened', 'onCardClose', 'onPageTabShow', 'onPageTabHide']);
   },
 
   mounted() {
@@ -258,6 +258,8 @@ export default {
       f7.on('pageNavbarLargeExpanded', self.onPageNavbarLargeExpanded);
       f7.on('cardOpened', self.onCardOpened);
       f7.on('cardClose', self.onCardClose);
+      f7.on('pageTabShow', self.onPageTabShow);
+      f7.on('pageTabHide', self.onPageTabHide);
     });
   },
 
@@ -284,6 +286,8 @@ export default {
     f7.off('pageNavbarLargeExpanded', self.onPageNavbarLargeExpanded);
     f7.off('cardOpened', self.onCardOpened);
     f7.off('cardClose', self.onCardClose);
+    f7.off('pageTabShow', self.onPageTabShow);
+    f7.off('pageTabHide', self.onPageTabHide);
     self.eventTargetEl = null;
     delete self.eventTargetEl;
   },
@@ -481,6 +485,16 @@ export default {
       this.setState({
         hasCardExpandableOpened: false
       });
+    },
+
+    onPageTabShow(pageEl) {
+      if (this.eventTargetEl !== pageEl) return;
+      this.dispatchEvent('page:tabshow pageTabShow');
+    },
+
+    onPageTabHide(pageEl) {
+      if (this.eventTargetEl !== pageEl) return;
+      this.dispatchEvent('page:tabhide pageTabHide');
     },
 
     dispatchEvent(events, ...args) {